+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Version 0.1.1 | 9/29/2021
+- Added `/pause` command
+- Added `/resume` command
+- Added `/stop` command
+- Added `/skip` command
+- Added `/help` command
+
 ## Version 0.1.0 | 9/28/2021
 - Added `/play` command
 - Change log started